requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.7
#django-heroku==0.3.1
django-timezone-field==4.0
djangorestframework==3.11.0
Faker==4.1.1
gunicorn==20.0.4
psycopg2-binary
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
whitenoise==5.1.0
requests==2.22.0
